commit_03

+++ controllers/contacts.js
@@ -1,10 +1,24 @@
 // controllers\contacts.js
 const service = require("../services/contacts");
 
-const listContacts = async (req, res) => {
+const getContactOwner = async (req, res) => {
   try {
-    // console.log(req.query);
-    const { success, result, message } = await service.listContacts();
+    // console.log(req.params.owner);
+    const owner = req.params.owner;
+    const { skip, limit, favorite } = req.query;
+    // console.log("test: ", req.query);
+
+    const query = { owner };
+    if (favorite !== undefined) {
+      query.favorite = favorite;
+    }
+
+    const { success, result, message } = await service.getContactOwner(
+      query,
+      skip,
+      limit
+    );
+
     if (!success) {
       return res.status(400).json({
         result,
@@ -24,13 +38,45 @@ const listContacts = async (req, res) => {
   }
 };
 
-const getContactById = async (req, res) => {
+const getContactOwnerById = async (req, res) => {
   try {
-    console.log(req.params.id);
-    const id = req.params.id;
-    const { success, result, message } = await service.getContactById(id);
+    // console.log(req.params.owner);
+    const { owner, id } = req.params;
+    // console.log("test: ", req.query);
+
+    const { success, result, message } = await service.getContactOwnerById(
+      owner,
+      id
+    );
+
     if (!success) {
       return res.status(400).json({
+        result,
+        message,
+      });
+    }
+
+    return res.status(200).json({
+      result,
+      message,
+    });
+  } catch (error) {
+    return res.status(500).json({
+      result: null,
+      message: error,
+    });
+  }
+};
+
+const getContactCurrent = async (req, res) => {
+  try {
+    // console.log(req.params.owner);
+    const id = req.user;
+
+    const { success, result, message } = await service.getContactCurrent(id);
+
+    if (!success) {
+      return res.status(401).json({
         result,
         message,
       });
@@ -50,9 +96,9 @@ const getContactById = async (req, res) => {
 
 const removeContact = async (req, res) => {
   try {
-    console.log(req.params.id);
-    const id = req.params.id;
-    const { success, result, message } = await service.removeContact(id);
+    // console.log(req.params.id);
+    const { owner, id } = req.params;
+    const { success, result, message } = await service.removeContact(owner, id);
     if (!success) {
       return res.status(400).json({
         result,
@@ -72,39 +118,28 @@ const removeContact = async (req, res) => {
   }
 };
 
-const addContact = async (req, res) => {
-  try {
-    const { success, result, message } = await service.addContact(req.body);
-    console.log(result);
-
-    if (!success) {
-      return res.status(400).json({
-        result,
-        message,
-      });
-    }
-
-    return res.status(201).json({
-      result,
-      message,
-    });
-  } catch (error) {
-    return res.status(500).json({
-      result: null,
-      message: error,
-    });
-  }
-};
-
 const updateContact = async (req, res) => {
   try {
-    const { id } = req.params;
+    const { owner, id } = req.params;
+    const { subscription } = req.body;
+    // console.log("sub ", subscription);
+    // console.log("sub1 ", id);
+
+    if (
+      subscription &&
+      !["starter", "pro", "business"].includes(subscription)
+    ) {
+      return res
+        .status(400)
+        .json({ result: null, message: "Ïnvalid subscription value" });
+    }
 
     const { success, result, message } = await service.updateContact(
+      owner,
       id,
-      req.body
+      subscription
     );
-    console.log(result);
+    // console.log("1 ", result);
 
     if (!success) {
       return res.status(400).json({
@@ -127,15 +162,17 @@ const updateContact = async (req, res) => {
 
 const updateStatusContact = async (req, res) => {
   try {
-    const { contactId } = req.params;
+    const { owner, id } = req.params;
     const { favorite } = req.body;
+    // console.log('id; ', id);
 
     const { success, result, message } = await service.updateStatusContact(
-      contactId,
+      owner,
+      id,
       favorite
     );
 
-    console.log(result);
+    // console.log("cons ", result);
 
     if (favorite === undefined) {
       return res.status(400).json({
@@ -163,11 +200,53 @@ const updateStatusContact = async (req, res) => {
   }
 };
 
+const updateTokenRemove = async (req, res) => {
+  try {
+    const { id } = req.params;
+    // console.log('contID: ', id);
+
+    const token = null;
+
+    const { success, result, message } = await service.updateTokenRemove(
+      id,
+      token
+    );
+
+    // console.log(result);
+
+    // if (token === undefined) {
+    //   return res.status(400).json({
+    //     result,
+    //     message: "missing field favorite",
+    //   });
+    // }
+
+    if (!success) {
+      return res.status(401).json({
+        result,
+        message,
+      });
+    }
+
+    return res.status(204).json({
+      result,
+      message,
+    });
+  } catch (error) {
+    return res.status(500).json({
+      result: null,
+      message: error,
+    });
+  }
+};
+
 module.exports = {
-  listContacts,
-  getContactById,
+  getContactOwner,
+  getContactOwnerById,
+  getContactCurrent,
   removeContact,
-  addContact,
+  // addContact,
   updateContact,
+  updateTokenRemove,
   updateStatusContact,
 };
